Add support for histogram pulse across UV
Inspired by AudioReactiveSurface and lilToon, but with the ability to
represent all 4 bands in one material. Still lacks an equivalent to
the "Delay" option in AudioReactiveSurface, however.

Maybe we could do away with the basic blink/beat option with this,
since setting ALBandXPulseScale to 0.0 achieves the same thing (this
is how things work in AudioReactiveSurface, for instance). Although
having both does allow for more potential effect combinations. Say
blinking on bass beat as a background to a bass beat histogram
scrolling across UV.

TODO:
  * We could do with starting to make the custom inspector less of a
    mess now that we have so many parameters. I still have just added
    the bare minimum so I can experiment.
  * Want to implement something similar for the UV shake/distortion
    options. Like a wave travelling through your UV warping it.